refactor for dashboard laboratory
