+++ frontend/package.json
@@ -31,6 +31,7 @@
     "react-dom": "^18.2.0",
     "react-qr-scanner": "^1.0.0-alpha.11",
     "react-scroll": "^1.9.0",
+    "react-toastify": "^9.1.3",
     "sharp": "^0.32.6",
     "viem": "~0.3.36",
     "wagmi": "^1.0.6",
