+++ frontend/src/app/layout.tsx
@@ -1,9 +1,11 @@
 "use client";
 import "./globals.css";
 import "./animation.css";
+import 'react-toastify/dist/ReactToastify.css';
 import { Providers } from "./providers";
 import Navbar from "../components/NavBar";
 import Footer from "../components/Footer";
+import { ToastContainer } from 'react-toastify';
 
 export default function RootLayout({
   children
@@ -14,6 +16,7 @@ export default function RootLayout({
     <html lang="en">
       <body className='body'>
         <Providers>
+          <ToastContainer />
           <Navbar />
           <div className='main'>
             {children}
