+++ frontend/src/components/Modal.tsx
@@ -1,5 +1,7 @@
 import React, { Fragment } from "react";
 import { Dialog, Transition } from "@headlessui/react";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faClose } from "@fortawesome/free-solid-svg-icons";
 
 const Modal = ({ isOpen, closeModal, children }: any) => {
   return (
@@ -28,7 +30,16 @@ const Modal = ({ isOpen, closeModal, children }: any) => {
               leaveFrom="opacity-100 scale-100"
               leaveTo="opacity-0 scale-95"
             >
-              <Dialog.Panel className="w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
+              <Dialog.Panel className="relative w-full max-w-md transform rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
+
+                <div style={{ width: '25px', height: '25px' }}
+                  onClick={closeModal}
+                  className="valign-middle cursor-pointer bg-slate-300 hover:bg-slate-200 absolute rounded-full right-5 top-5 ta-c">
+                  <div>
+                    <FontAwesomeIcon className="text-slate-500 text-sm " icon={faClose} />
+                  </div>
+                </div>
+
                 {children}
               </Dialog.Panel>
             </Transition.Child>
